fix: moved modal tracker inside modal
